refact: move lib to server
