+++ server/embed.js
